update nuget publish scripts

+++ CreateLocalPackages.ps1
@@ -1,15 +1,18 @@
 $scriptPath = Split-Path $MyInvocation.MyCommand.Path
 
+Write-Output $scriptPath
+
 Set-Location $scriptPath
 
-. ./BuildAll.ps1
+#. ./BuildAll.ps1
 
 . ./UpdateNuspecVersion.ps1 -preRelease alpha -major 0 -minor 8
 
 $oldChildren = gci $scriptPath -Recurse *.nupkg
-foreach( $oldChild in $oldChildren)
-{
-	Remove-Item $oldChild -Recurse
+foreach( $oldChild in $oldChildren){
+	if(!($oldChild.DirectoryName -match "packages")){
+		Remove-Item $oldChild -Recurse
+	}
 }
 
 $children = gci $scriptPath -recurse *.nuspec
